+++ changelog.md
@@ -5,3 +5,7 @@ Harvest [bridge-adapter] (2019-09-06)
 Harvest [bridge-adapter] (2019-12-31)
   * \[kinetic-bridge-adapter-harvest]
     * All unit tests passed or skipped.
+
+Harvest [bridge-adapter] (2020-06-09)
+  * \[kinetic-bridge-adapter-harvest] [v2.1.0]
+    * Refactored the adapter to leverage a different "adapter api".